+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.5.4 (Next)
 
+* Enable setting `send_gifs` per instance of `SlackRubyBot::Server` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.5.3 (12/28/2015)